% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadepend_prior.R
\name{data_dep_prior_multi}
\alias{data_dep_prior_multi}
\title{Parameters of the data-dependent prior distribution}
\usage{
data_dep_prior_multi(dat)
}
\arguments{
\item{dat}{dataset}
}
\value{
The original data will have one more column at the end, which is the parameter of the prior distribution.
}
\description{
Creates a data-dependent prior for a p-dimensional multinomial distribution. Only referenced if conj_prior is specified.
Use a conjugate prior based on 20% of the data (e.g. \eqn{Dirichlet(\alpha)}).
}
