% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkprior.R
\name{check_prior}
\alias{check_prior}
\title{Check Prior Distribution Parameters}
\usage{
check_prior(
  dat,
  conj_prior = c("none", "data.dep", "flat.prior", "non.informative"),
  alpha = NULL,
  outer = FALSE,
  enum_comp = NULL
)
}
\arguments{
\item{dat}{A dataset (usually of class `data.frame` or `data.table`) containing the data used to calculate the prior or posterior parameters.}

\item{conj_prior}{A string indicating the type of prior distribution to use. Can be one of:
- "none": No prior distribution, generate random theta values.
- "data.dep": A data-dependent prior (user-supplied or calculated from `dat`).
- "flat.prior": A flat prior with a scalar alpha parameter.
- "non.informative": A non-informative Jeffrey's prior (alpha = 1).}

\item{alpha}{A scalar or vector that defines the prior distribution. The exact behavior depends on `conj_prior`:
- For "flat.prior", this should be a scalar.
- For "data.dep", this will be either user-supplied or calculated from the data.
- For "non.informative", this is ignored and set to 1.}

\item{outer}{A logical value (default is `FALSE`). If `TRUE`, the function is called within the `multinomial_impute` function, which controls how prior and posterior distributions are handled.}

\item{enum_comp}{A `data.table` or `data.frame` that stores enumeration components, which will be updated with prior/posterior parameters.}
}
\value{
If `outer` is `TRUE`, returns the prior (alpha) as a scalar or vector, depending on the type of prior selected.
        If `outer` is `FALSE`, the function returns `enum_comp` with two additional columns:
        - The parameters of the prior distribution (`alpha` or calculated values).
        - The parameters of the posterior distribution (`theta_y`).
}
\description{
Check Prior Distribution Parameters
}
\section{Functions}{
\itemize{
\item \code{check_prior()}: the parameters of the prior distribution OR
generate the posterior distribution from the parameters of the prior distribution.

}}
