% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getleveltext.R
\name{get_level_text}
\alias{get_level_text}
\title{Get Level Text}
\usage{
get_level_text(var, val)
}
\arguments{
\item{var}{Name of the categorical variable.}

\item{val}{The positive integers converted to categorical variable factor levels.}
}
\value{
Text form of factor levels for categorical variables.
}
\description{
Get Level Text
}
