% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute4mmv.R
\name{impute4mmv}
\alias{impute4mmv}
\title{Impute Missing Values for Multinomial Data}
\usage{
impute4mmv(
  dat,
  method = c("EM", "DA"),
  conj_prior = c("none", "data.dep", "flat.prior", "non.informative"),
  alpha = NULL
)
}
\arguments{
\item{dat}{A data.table or data frame containing the data with missing values.
All columns should be factors.}

\item{method}{A character string specifying the imputation method.
Options include "EM" for Expectation-Maximization and "DA" for Dirichlet Allocation.}

\item{conj_prior}{A character string specifying the type of prior to use.
Options include "none", "data.dep", "flat.prior", and "non.informative".}

\item{alpha}{A numeric vector specifying the prior parameters. Required if
`conj_prior` is "flat.prior".}
}
\value{
An object of class `imputeMulti` containing:
\item{Gcall}{The call to the function.}
\item{method}{The method used for imputation.}
\item{mle_call}{The call used for maximum likelihood estimation.}
\item{mle_iter}{The number of iterations in MLE.}
\item{mle_log_lik}{Log-likelihood of the fitted model.}
\item{mle_cp}{Conjugate prior used.}
\item{mle_x_y}{Estimated parameters for the multinomial distribution.}
\item{data}{A list containing the original missing data and the imputed data.}
\item{nmiss}{The number of missing observations.}
}
\description{
This function imputes missing values in categorical data using either
Expectation-Maximization (EM) or Dirichlet Allocation (DA) methods.
It utilizes prior distributions for the imputation process.
}
