% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute1.R
\name{impute_multinomial}
\alias{impute_multinomial}
\title{Impute Missing Values in Multinomial Data}
\usage{
impute_multinomial(miss_val, row_ind, MLEx_y)
}
\arguments{
\item{miss_val}{A data frame or vector containing missing values to be imputed.}

\item{row_ind}{An integer index indicating the row in `MLEx_y` to use for imputation.}

\item{MLEx_y}{A data frame containing maximum likelihood estimates for imputation.}
}
\value{
A data frame with missing values imputed.
}
\description{
This function imputes missing values in a single row of multinomial data using
maximum likelihood estimates based on the provided `MLEx_y` data.
}
