% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute2.R
\name{impute_multinomial_all}
\alias{impute_multinomial_all}
\title{Impute Missing Values Across All Rows in Multinomial Data

This function imputes missing values for all rows in a dataset using the
`impute_multinomial` function based on the provided `MLEx_y` data.}
\usage{
impute_multinomial_all(dat_miss, MLEx_y, p)
}
\arguments{
\item{dat_miss}{A data frame with missing values to be imputed.}

\item{MLEx_y}{A data frame containing maximum likelihood estimates for imputation.}

\item{p}{An integer specifying the number of columns in `MLEx_y` to consider.}
}
\value{
A data frame with all missing values imputed.
}
\description{
Impute Missing Values Across All Rows in Multinomial Data

This function imputes missing values for all rows in a dataset using the
`impute_multinomial` function based on the provided `MLEx_y` data.
}
