% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkClass2.R
\docType{class}
\name{imputeMulti-class}
\alias{imputeMulti-class}
\title{Make A Class}
\description{
Make A Class
}
\section{Slots}{

\describe{
\item{\code{Gcall}}{call.}

\item{\code{method}}{character.}

\item{\code{mle_call}}{call.}

\item{\code{mle_iter}}{numeric.}

\item{\code{mle_log_lik}}{numeric.}

\item{\code{mle_cp}}{character.}

\item{\code{mle_x_y}}{data.frame.}

\item{\code{data}}{list.}

\item{\code{nmiss}}{numeric.}
}}

