% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isClass2.R
\name{is.imputeMulti}
\alias{is.imputeMulti}
\title{Function that checks if the target object is a \code{mod_imputeMulti} object.}
\usage{
is.imputeMulti(x)
}
\arguments{
\item{x}{}
}
\value{
TRUE OR FALSE
}
\description{
Function that checks if the target object is a \code{mod_imputeMulti} object.
}
