% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkClass1.R
\docType{class}
\name{mod_imputeMulti-class}
\alias{mod_imputeMulti-class}
\title{Making a Class}
\description{
Making a Class
}
\section{Slots}{

\describe{
\item{\code{method}}{character.}

\item{\code{mle_call}}{call.}

\item{\code{mle_iter}}{numeric.}

\item{\code{mle_log_lik}}{numeric.}

\item{\code{mle_cp}}{character.}

\item{\code{mle_x_y}}{data.frame.}
}}

