% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{EM_multinomial}
\alias{EM_multinomial}
\title{Expectation-Maximization Algorithm for Multinomial Data}
\usage{
EM_multinomial(
  dat,
  conj_prior = c("none", "data.dep", "flat.prior", "non.informative"),
  alpha = NULL,
  tol = 5e-07,
  max_iter = 10000
)
}
\arguments{
\item{dat}{A data frame containing the data with missing values. All columns
should be factors.}

\item{conj_prior}{A character string specifying the type of prior to use.
Options include "none", "data.dep", "flat.prior", and "non.informative".}

\item{alpha}{A numeric vector specifying the prior parameters. Required if
`conj_prior` is "data.dep" or "flat.prior".}

\item{tol}{A numeric value indicating the convergence tolerance. Default is
5e-7.}

\item{max_iter}{An integer specifying the maximum number of iterations.
Default is 10000.}
}
\value{
An object of class `mod_imputeMulti` containing:
\item{method}{The method used for imputation.}
\item{mle_call}{The call that produced the model.}
\item{mle_iter}{The number of iterations run.}
\item{mle_log_lik}{The log-likelihood of the fitted model.}
\item{mle_cp}{The conjugate prior used.}
\item{mle_x_y}{A data frame with estimated parameters and counts.}
}
\description{
This function implements the Expectation-Maximization (EM) algorithm to
estimate parameters for multinomial data with missing values.
}
