% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2int.R
\name{fact_to_int}
\alias{fact_to_int}
\title{Convert the factor levels of categorical variables to positive integers.}
\usage{
fact_to_int(f)
}
\arguments{
\item{f}{A column of factor variables}
}
\value{
Factor levels converted to a column of positive integers.
}
\description{
Convert the factor levels of categorical variables to positive integers.
}
