% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isClass2.R
\name{is.imputeMulti}
\alias{is.imputeMulti}
\title{Check if the target object is a \code{imputeMulti} object}
\usage{
is.imputeMulti(x)
}
\arguments{
\item{x}{The object to be checked.}
}
\value{
A logical value: \code{TRUE} if \code{x} is an \code{imputeMulti} object, \code{FALSE} otherwise.
}
\description{
This function checks whether the provided object is of class \code{imputeMulti}.
}
