% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isClass.R
\name{is.mod_imputeMulti}
\alias{is.mod_imputeMulti}
\title{Check if the target object is a \code{mod_imputeMulti} object}
\usage{
is.mod_imputeMulti(x)
}
\arguments{
\item{x}{The object to be checked.}
}
\value{
A logical value: \code{TRUE} if \code{x} is a \code{mod_imputeMulti} object, \code{FALSE} otherwise.
}
\description{
This function checks whether the provided object is of class \code{mod_imputeMulti}.
}
